requirements.txt
anonfile
discord.py-self
easygui
python-dotenv